6th Commit
In this commit I added the area cropping NASS cdl, where it inserts GIS processing to the calculation of the land river segments. It also adds some files that are still under progress such as the AFIA_feed_allocation that deals with new diet assumptions to the model.
There are also some updates in the coding of the grass scenario (new values for constants). Added new numbers for dairy production, and diet. Recalculated population distribution using GIS information for LRS. Changed the number of counties to 202. Changed meatdata_master, cropfertrates, cropdata master and animdatadyn master,